+++ src/apps/ReleaseControl/components/Form/Inputs/Scope/TagTable.tsx
@@ -1,12 +1,11 @@
 import { tokens } from '@equinor/eds-tokens';
-import { proCoSysUrls, stidUrls } from '@equinor/procosys-urls';
+import { proCoSysUrls, stidUrls, echoUrls } from '@equinor/procosys-urls';
 import { CellProps, Column, Table, defaultGroupByFn } from '@equinor/Table';
 import { FamTagType } from '@equinor/Workflow';
 import styled from 'styled-components';
 import { RemoveTagCell } from './RemoveTagCell';
 import { Icon } from '@equinor/eds-core-react';
 import { LinkGroup } from './LinkGroup';
-import { Echo3DIconLink } from './Echo3DIconLink';
 interface TagTableProps {
     tags: FamTagType[];
     editMode: boolean;
@@ -37,25 +36,35 @@ const columns: Column<FamTagType>[] = [
                 >
                     {cell.row.values.tagNo}
                 </Link>
-                <Echo3DIconLink id={cell.row.original.tagNo} />
             </div>
         ),
         Aggregated: () => null,
         aggregate: 'count',
     },
     {
-        id: 'stidLink',
+        id: 'links',
         Header: 'Links',
-        width: 60,
+        width: 70,
         accessor: (item) => ({
             content: item,
             currentKey: 'tagNo',
-            url: stidUrls.getTagUrl(item.tagNo),
+            stidUrl: stidUrls.getTagUrl(item.tagNo),
+            echoUrl: echoUrls.getEchoUrl(item.tagNo),
         }),
         Cell: (cell: CellProps<FamTagType>) => (
-            <Link href={cell.value.url} target="_blank" hideUnderline>
-                <StidLogoLink src="images/stid_logo.svg" />
-            </Link>
+            <StyledLinkGrouping>
+                <Link href={cell.value.stidUrl} target="_blank" hideUnderline title="Open in STID">
+                    <StyledStidLogoLink src="images/stid_logo.svg" />
+                </Link>
+                <Link
+                    href={cell.value.echoUrl}
+                    target="_blank"
+                    hideUnderline
+                    title="Open in Echo 3D"
+                >
+                    <StyledEchoLogoLink src="images/echo_logo.svg" />
+                </Link>
+            </StyledLinkGrouping>
         ),
         Aggregated: () => null,
         aggregate: 'count',
@@ -66,16 +75,16 @@ const columns: Column<FamTagType>[] = [
         accessor: (item) => item.register,
     },
     {
-        id: 'tagMountedOnNo',
+        id: 'tagMountedOn',
         Header: 'Mounted on',
         accessor: (item) => ({
             content: item,
-            currentKey: 'tagMountedOnNo',
+            currentKey: 'tagMountedOn',
             url: proCoSysUrls.getTagUrl(item.tagMountedOnUrlId || ''),
         }),
         Cell: (cell: CellProps<FamTagType>) => (
             <Link href={cell.value.url} target="_blank" hideUnderline>
-                {cell.value.content.tagMountedOnNo}
+                {cell.value.content.tagMountedOn}
             </Link>
         ),
         Aggregated: () => null,
@@ -209,6 +218,16 @@ const Link = styled.a`
     }
 `;
 
-const StidLogoLink = styled.img`
+const StyledStidLogoLink = styled.img`
     width: 24px;
+`;
+const StyledEchoLogoLink = styled.img`
+    width: 20px;
+`;
+
+const StyledLinkGrouping = styled.div`
+    display: flex;
+    align-items: center;
+    gap: 0.2em;
+    justify-content: space-evenly;
 `;